Cleaned up code structure and added a function
feel free to let me know how much u hate me
