+++ .gitignore
@@ -1,0 +1,4 @@
+
+/Projects
+/__pycache__
+/BRCI_RF/__pycache__
