+++ BRCI_RF/__init__.py
@@ -1,2 +1,2 @@
-from .functions import *
 from .brick_list import *
+from .functions import *
